feat(frontend): mysql、tendbcluster定点构造改造 #5611 # Reviewed, transaction id: 13523
