+++ dbm-ui/frontend/src/views/mysql/rollback/pages/page1/components/render-data/exist-cluster/Row.vue
@@ -86,8 +86,8 @@
   import RenderCluster from '@views/mysql/rollback/pages/page1/components/common/RenderCluster.vue';
   import RenderClusterInputSelect from '@views/mysql/rollback/pages/page1/components/common/RenderClusterInputSelect.vue';
 
-  import { createRowData, type IDataRow } from '../../../Index.vue';
   import { BackupSources } from '../../common/const';
+  import { createRowData, type IDataRow } from '../Index.vue';
 
   interface Props {
     data: IDataRow;
@@ -164,6 +164,7 @@
     },
     {
       immediate: true,
+      deep: true,
     },
   );
