+++ dbm-ui/frontend/src/views/spider-manage/rollback/pages/page1/components/render-data/exist-cluster/Row.vue
@@ -88,9 +88,9 @@
   import RenderBackup from '@views/mysql/rollback/pages/page1/components/common/RenderBackup.vue';
   import RenderCluster from '@views/mysql/rollback/pages/page1/components/common/RenderCluster.vue';
 
-  import { createRowData, type IDataRow } from '../../../Index.vue';
   import { BackupSources, selectList } from '../../common/const';
   import RenderClusterInputSelect from '../../common/RenderClusterInputSelect.vue';
+  import { createRowData, type IDataRow } from '../Index.vue';
 
   interface Props {
     data: IDataRow;
@@ -167,6 +167,7 @@
     },
     {
       immediate: true,
+      deep: true,
     },
   );
