+++ dbm-ui/frontend/src/views/spider-manage/rollback/pages/page1/components/render-data/new-cluster/Index.vue
@@ -219,11 +219,11 @@
   import RenderTableHeadColumn from '@components/render-table/HeadColumn.vue';
   import RenderTable from '@components/render-table/Index.vue';
 
-  import type { IDataRow } from '../../../Index.vue';
   import { BackupTypes, selectList } from '../../common/const';
+  import type { IDataRow } from '../Index.vue';
 
   interface Emits {
-    (e: 'batchSelectCluster'): void;
+    (e: 'showSelector'): void;
     (e: 'batchEdit', obj: Record<string, any>): void;
   }
 
@@ -263,7 +263,7 @@
   };
 
   const handleShowBatchSelector = () => {
-    emits('batchSelectCluster');
+    emits('showSelector');
   };
 
   const handleBatchEdit = (key: keyof IDataRow, value: string, isArray = false) => {
