+++ dbm-ui/frontend/src/views/spider-manage/rollback/pages/page1/components/render-data/origin-cluster/Row.vue
@@ -52,8 +52,8 @@
   import RenderBackup from '@views/mysql/rollback/pages/page1/components/common/RenderBackup.vue';
   import RenderCluster from '@views/mysql/rollback/pages/page1/components/common/RenderCluster.vue';
 
-  import { createRowData, type IDataRow } from '../../../Index.vue';
   import { BackupSources, selectList } from '../../common/const';
+  import { createRowData, type IDataRow } from '../Index.vue';
 
   interface Props {
     data: IDataRow;
@@ -125,6 +125,7 @@
     },
     {
       immediate: true,
+      deep: true,
     },
   );
